+++ src/Ext/TechnoType/Body.h
@@ -429,6 +429,9 @@ public:
 	Valueable<double> Convert_Health_BelowPercent;
 	Nullable<TechnoTypeClass*> Convert_Health;
 
+	ValueableVector<int> DefaultToGuardArea_Modes;
+	ValueableVector<int> DefaultToGuardArea_AIModes;
+
 	// Ares 0.2
 	Valueable<int> RadarJamRadius;
 
@@ -829,6 +832,9 @@ public:
 		, Convert_Health_AbovePercent { -1.0 }
 		, Convert_Health_BelowPercent { -1.0 }
 		, Convert_Health {}
+			
+		, DefaultToGuardArea_Modes {}
+		, DefaultToGuardArea_AIModes {}
 
 		// Ares 0.2
 		, RadarJamRadius { 0 }